Moving whole Plausible logic to useTracking hook, that is watching whether the exercise has just started or finished. However when opened from dictionary, it is logged into the console, as well as when the exercise is finished - yet to fix
