+++ components/basecomponents/Exercise/ExerciseConfiguration.tsx
@@ -5,7 +5,7 @@ import { useDebug } from './utils/useDebug';
 import { ActionButton } from './components/ActionButton';
 import { useTranslation } from 'react-i18next';
 import { CONFIG_BASE } from './exerciseStoreConfig';
-import { useTracking } from 'utils/useTracking';
+import { useLanguage } from 'utils/useLanguageHook';
 
 const removeDuplicates = (array: string[]) => {
   return [...new Set(array)];
@@ -19,12 +19,11 @@ const removeDuplicates = (array: string[]) => {
  * Exercise store only deals with categories, all complexity of mapping metacategories to categories is handled here.
  */
 const ExerciseConfiguration: FunctionComponent = () => {
-  const { lang, plausible } = useTracking();
-
   // Categories passed  to Exercise store to be used for generating exercises
   const setCategories = useExerciseStore((state) => state.setCategories);
   const setSize = useExerciseStore((state) => state.setSize);
   const size = useExerciseStore((state) => state.size);
+  const lang = useLanguage();
 
   // User-selected metacategories, each metacategory contains multiple categories
   const [selectedMetaIds, setSelectedMetaIds] = useState<string[]>([]);
@@ -82,13 +81,7 @@ const ExerciseConfiguration: FunctionComponent = () => {
         </div>
       </div>
       <div className="flex flex-col items-center mb-12">
-        <ActionButton
-          onClick={() => {
-            plausible('Exercise-Started', { props: { language: lang.currentLanguage, length: size } });
-          }}
-          action="start"
-          disabled={selectedMetaIds.length === 0}
-        />
+        <ActionButton action="start" disabled={selectedMetaIds.length === 0} />
       </div>
     </div>
   );
