+++ pages/dictionary/index.tsx
@@ -16,7 +16,7 @@ import { getServerSideTranslations } from '../../utils/localization';
 import { TextLink } from '../../components/Typography';
 import { AiOutlineFilePdf } from 'react-icons/ai';
 import { getCategoryName, getCategoryId, getCategoryUkId } from '../../components/sections/Dictionary/dictionaryUtils';
-import { useTracking } from 'utils/useTracking';
+import { useLanguage } from 'utils/useLanguageHook';
 
 // Disable ssr for this component to avoid Reference Error: Blob is not defined
 const ExportTranslations = dynamic(() => import('../../components/sections/ExportTranslations'), {
@@ -26,7 +26,7 @@ const ExportTranslations = dynamic(() => import('../../components/sections/Expor
 const Dictionary = ({ dictionary }: InferGetStaticPropsType<typeof getStaticProps>) => {
   const categories = useMemo(() => getCategories(dictionary), [dictionary]);
   const allTranslations = useMemo(() => getAllPhrases(dictionary), [dictionary]);
-  const { lang, plausible } = useTracking();
+  const lang = useLanguage();
 
   const [search, setSearch] = useState('');
   const [isSticky, setIsSticky] = useState(false);
@@ -150,13 +150,7 @@ const Dictionary = ({ dictionary }: InferGetStaticPropsType<typeof getStaticProp
                         locale={getCountryVariant()}
                       >
                         <AiOutlineFilePdf className="w-5 h-5" />
-                        <p
-                          onClick={() => {
-                            plausible('Exercise-Started', { props: { language: lang.currentLanguage, length: 10 } });
-                          }}
-                        >
-                          {t('header.exercises_link_name')}
-                        </p>
+                        {t('header.exercises_link_name')}
                       </TextLink>
                     </div>
                     <CategoryDictionary searchText={search} translations={category.translations} />
